+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Changed
 - ServiceClientFactory deprecated by VertxEndpointClient
+- ComponentVerticle does not print stacktrace when TracingVerticle not available
 
 ## [1.1.0] - 2019-04-20
 ### Added